In 2nd commit I have addressed initial JanK's feedback. I have fixed visibility specifiers to go first everywhere. I have removed the file Relocator.cs; instead I created a new file RelocationHelper.cs where I moved the entire class RelocationHelper previously residing within the SectionBuilder class.
I have removed the individual relocators and replaced them with a plain
switch per Jan's suggestion. I have also switched the relocation process
to use a common buffer to avoid repeatedly allocating.

I have also found out that the pre-existing method of section
composition was pretty wasteful - as each block was holding onto its
own content BlobBuilder, with a large number of blocks the BlobBuilder
overhead could become huge.

I have refactored this part so that the subsections are getting laid
out immediately as blocks are added i.e. blocks no longer hold onto
their content and instead we create a [typically large] subsection
content BlobBuilder which is much more efficient.

This change also fixes two bugs from the first commit uncovered in
local testing:

1) We must ensure that each serialized section has at least 1 byte,
otherwise the PEBuilder goes mad and corrupts the output executable;

2) We must reset the ILOnly bit flag on CorHeader.Flags, otherwise
Windows PE loader doesn't relocate the file and marks all pages as
read-only.

With these additional changes, I have a working "mini-demo" where
I can use a native app to LoadLibrary the DLL produced by the
R2RPEBuilder, call a native export on it and print its return value
which is a char * in the data segment (i.e. this demonstrates that
at least the HIGHLOW file-level relocation works).

I have also created two more "unit tests" (I'm not yet proposing
their submission as it will need a bit of work to plug them into
the XUnit test system) - TestSizeStress (which creates a 200 MB
PE using 4 MB 50-byte blocks) and TestRelocationStress (which creates
a 200M section with 50M relocations).

Before the subsection content optimization change, the TestSizeStress
used to grow GC heap by about 2 GB while producing the 200 MB PE file.
After the change GC heap growth went down to about 600 MB. Duration
of the stress test (in x86 release build) is about 3 seconds.

For the TestRelocationStress, GC heap growth is about 2 GB (please
note that, for 100M relocations, 1600 MB is just the raw size of
the Relocation structs which have 16 bytes each) and in the release
build it also takes about 3 seconds.

Thanks

Tomas
